nq/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=43.0.0", "wheel"]

[project]
dependencies = [
  "tqdm",
  "numpy",
  "seaborn",
  "scikit-learn",
  "matplotlib",
  "ipython",
  "jupyter",
  "quapy",
  "graphesn",
  "torch==2.6.0",
  "torch_geometric",
  "python-dotenv",
  "rootutils",
  "hydra-core>=1.3.1",
  "hydra-colorlog>=1.2.0",
  "rich",
  "omegaconf",
  "demon",
]
name = "nq"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.scripts]
nq-quant = "nq.quantify:main"
nq-train = "nq.train:main"

[tool.setuptools.packages.find]
include = ["nq*"] # package names should match these glob patterns (["*"] by default)
where = ["src"] # list of folders that contain the packages (["."] by default)

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
